+++ squadcastv1/src/funcs/statusPagesList.ts
@@ -3,8 +3,8 @@
  */
 
 import { SquadcastSDKCore } from "../core.js";
-import { dlv } from "../lib/dlv.js";
 import { encodeFormQuery } from "../lib/encodings.js";
+import { matchStatusCode } from "../lib/http.js";
 import * as M from "../lib/matchers.js";
 import { compactMap } from "../lib/primitives.js";
 import { safeParse } from "../lib/schemas.js";
@@ -171,21 +171,8 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: [
-      "400",
-      "401",
-      "402",
-      "403",
-      "404",
-      "409",
-      "422",
-      "4XX",
-      "500",
-      "502",
-      "503",
-      "504",
-      "5XX",
-    ],
+    isErrorStatusCode: (statusCode: number) =>
+      matchStatusCode({ status: statusCode } as Response, ["4XX", "5XX"]),
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -280,7 +267,7 @@ async function $do(
     if (!responseData) {
       return { next: () => null };
     }
-    const results = dlv(responseData, "data");
+    const results = (responseData as { data: unknown }).data;
     if (!Array.isArray(results) || !results.length) {
       return { next: () => null };
     }